Update README.md
The file /vagrant/puphpet/files/exec-once/config-nginx.sh is not executed at the moment of first vagrant up. What is the reason for this? At the moment I am not sure (maybe the bug of puphpet, maybe something else, cannot find it quickly...). I leave it as it is, because typing two additional lines into the shell would fix the problem of nginx configured incorrectly.